+++ src/commands/diffWithBranch.ts
@@ -5,7 +5,7 @@ import { GlyphChars } from '../constants';
 import { DiffWithCommandArgs } from './diffWith';
 import { GitService, GitUri } from '../gitService';
 import { Messages } from '../messages';
-import { BranchesQuickPick, CommandQuickPickItem } from '../quickPicks';
+import { BranchesAndTagsQuickPick, CommandQuickPickItem } from '../quickPicks';
 import * as path from 'path';
 
 export interface DiffWithBranchCommandArgs {
@@ -35,43 +35,57 @@ export class DiffWithBranchCommand extends ActiveEditorCommand {
         const gitUri = await GitUri.fromUri(uri, this.git);
         if (!gitUri.repoPath) return Messages.showNoRepositoryWarningMessage(`Unable to open branch compare`);
 
-        const branches = await this.git.getBranches(gitUri.repoPath);
-        const pick = await BranchesQuickPick.show(branches, `Compare ${path.basename(gitUri.fsPath)} to ${GlyphChars.Ellipsis}`, args.goBackCommand);
-        if (pick === undefined) return undefined;
+        const placeHolder = `Compare ${path.basename(gitUri.fsPath)} with ${GlyphChars.Ellipsis}`;
+        const progressCancellation = BranchesAndTagsQuickPick.showProgress(placeHolder);
 
-        if (pick instanceof CommandQuickPickItem) return pick.execute();
+        try {
+            const [branches, tags] = await Promise.all([
+                this.git.getBranches(gitUri.repoPath),
+                this.git.getTags(gitUri.repoPath)
+            ]);
 
-        const branch = pick.branch.name;
-        if (branch === undefined) return undefined;
+            if (progressCancellation.token.isCancellationRequested) return undefined;
 
-        let renamedUri: Uri | undefined;
-        let renamedTitle: string | undefined;
+            const pick = await BranchesAndTagsQuickPick.show(branches, tags, placeHolder, { progressCancellation: progressCancellation, goBackCommand: args.goBackCommand });
+            if (pick === undefined) return undefined;
 
-        // Check to see if this file has been renamed
-        const statuses = await this.git.getDiffStatus(gitUri.repoPath, 'HEAD', branch, { filter: 'R' });
-        if (statuses !== undefined) {
-            const fileName = GitService.normalizePath(path.relative(gitUri.repoPath, gitUri.fsPath));
-            const rename = statuses.find(s => s.fileName === fileName);
-            if (rename !== undefined && rename.originalFileName !== undefined) {
-                renamedUri = Uri.file(path.join(gitUri.repoPath, rename.originalFileName));
-                renamedTitle = `${path.basename(rename.originalFileName)} (${branch})`;
+            if (pick instanceof CommandQuickPickItem) return pick.execute();
+
+            const ref = pick.name;
+            if (ref === undefined) return undefined;
+
+            let renamedUri: Uri | undefined;
+            let renamedTitle: string | undefined;
+
+            // Check to see if this file has been renamed
+            const statuses = await this.git.getDiffStatus(gitUri.repoPath, 'HEAD', ref, { filter: 'R' });
+            if (statuses !== undefined) {
+                const fileName = GitService.normalizePath(path.relative(gitUri.repoPath, gitUri.fsPath));
+                const rename = statuses.find(s => s.fileName === fileName);
+                if (rename !== undefined && rename.originalFileName !== undefined) {
+                    renamedUri = Uri.file(path.join(gitUri.repoPath, rename.originalFileName));
+                    renamedTitle = `${path.basename(rename.originalFileName)} (${ref})`;
+                }
             }
-        }
 
-        const diffArgs: DiffWithCommandArgs = {
-            repoPath: gitUri.repoPath,
-            lhs: {
-                sha: pick.branch.remote ? `remotes/${branch}` : branch,
-                uri: renamedUri || gitUri as Uri,
-                title: renamedTitle || `${path.basename(gitUri.fsPath)} (${branch})`
-            },
-            rhs: {
-                sha: '',
-                uri: gitUri as Uri
-            },
-            line: args.line,
-            showOptions: args.showOptions
-        };
-        return commands.executeCommand(Commands.DiffWith, diffArgs);
+            const diffArgs: DiffWithCommandArgs = {
+                repoPath: gitUri.repoPath,
+                lhs: {
+                    sha: pick.remote ? `remotes/${ref}` : ref,
+                    uri: renamedUri || gitUri as Uri,
+                    title: renamedTitle || `${path.basename(gitUri.fsPath)} (${ref})`
+                },
+                rhs: {
+                    sha: '',
+                    uri: gitUri as Uri
+                },
+                line: args.line,
+                showOptions: args.showOptions
+            };
+            return commands.executeCommand(Commands.DiffWith, diffArgs);
+        }
+        finally {
+            progressCancellation.dispose();
+        }
     }
 }