+++ src/commands/diffWithRevision.ts
@@ -7,7 +7,7 @@ import { DiffWithCommandArgs } from './diffWith';
 import { GitService, GitUri } from '../gitService';
 import { Logger } from '../logger';
 import { Messages } from '../messages';
-import { CommandQuickPickItem, FileHistoryQuickPick } from '../quickPicks';
+import { CommandQuickPickItem, FileHistoryQuickPick, ShowBranchesAndTagsQuickPickItem } from '../quickPicks';
 
 export interface DiffWithRevisionCommandArgs {
     maxCount?: number;
@@ -36,7 +36,9 @@ export class DiffWithRevisionCommand extends ActiveEditorCommand {
 
         const gitUri = await GitUri.fromUri(uri, this.git);
 
-        const progressCancellation = FileHistoryQuickPick.showProgress(gitUri);
+        const placeHolder = `Compare ${gitUri.getFormattedPath()}${gitUri.sha ? ` ${Strings.pad(GlyphChars.Dot, 1, 1)} ${gitUri.shortSha}` : ''} with ${GlyphChars.Ellipsis}`;
+        const progressCancellation = FileHistoryQuickPick.showProgress(placeHolder);
+
         try {
             const log = await this.git.getLogForFile(gitUri.repoPath, gitUri.fsPath, { maxCount: args.maxCount, ref: gitUri.sha });
             if (log === undefined) return Messages.showFileNotUnderSourceControlWarningMessage('Unable to open history compare');
@@ -60,9 +62,13 @@ export class DiffWithRevisionCommand extends ActiveEditorCommand {
                 }
             }
 
-            const label = `${gitUri.getFormattedPath()}${gitUri.sha ? ` ${Strings.pad(GlyphChars.Dot, 1, 1)} ${gitUri.shortSha}` : ''}`;
-            const pick = await FileHistoryQuickPick.show(this.git, log, gitUri, label, progressCancellation, {
+            const pick = await FileHistoryQuickPick.show(this.git, log, gitUri, placeHolder, {
                 pickerOnly: true,
+                progressCancellation: progressCancellation,
+                currentCommand: new CommandQuickPickItem({
+                    label: `go back ${GlyphChars.ArrowBack}`,
+                    description: `${Strings.pad(GlyphChars.Dash, 2, 3)} to history of ${GlyphChars.Space}$(file-text) ${gitUri.getFormattedPath()}${gitUri.sha ? ` from ${GlyphChars.Space}$(git-commit) ${gitUri.shortSha}` : ''}`
+                }, Commands.DiffWithRevision, [uri, { ...args }]),
                 nextPageCommand: args.nextPageCommand,
                 previousPageCommand: previousPageCommand,
                 showAllCommand: log !== undefined && log.truncated
@@ -74,12 +80,26 @@ export class DiffWithRevisionCommand extends ActiveEditorCommand {
             });
             if (pick === undefined) return undefined;
 
-            if (pick instanceof CommandQuickPickItem) return pick.execute();
+            let ref: string;
+
+            if (pick instanceof ShowBranchesAndTagsQuickPickItem) {
+                const branchOrTag = await pick.execute();
+                if (branchOrTag === undefined) return undefined;
+
+                if (branchOrTag instanceof CommandQuickPickItem) return branchOrTag.execute();
+
+                ref = branchOrTag.name;
+            }
+            else {
+                if (pick instanceof CommandQuickPickItem) return pick.execute();
+
+                ref = pick.commit.sha;
+            }
 
             const diffArgs: DiffWithCommandArgs = {
                 repoPath: gitUri.repoPath,
                 lhs: {
-                    sha: pick.commit.sha,
+                    sha: ref,
                     uri: gitUri as Uri
                 },
                 rhs: {